Create a distribution structure
The basic type around which the various informational metrics are
constructed is the probability distribution, inform_distribution.
The underlying data structure is that of a histogram containing
frequencies (of type uint64_t) of events (also of type uint64_t).
A counter is used to keep track of the number of events observed
allowing for "conversion" from a frequency distribution to a
probability distribution.