Remove the state encoding code.
This has been coming for a while. It is not being used anywhere in
the code base. In the event that we need such functionality in the
future, it can be added back in.

It is likely that a coming version will include a "utilities"
section for things like state binning. State encoding may fit well
there.