Reduce AI calls to accumulate_observations
The previous implementation called accumulate_observations once
for each initial condition. This is likely not a problem given
that accumulate_observations is a static function; it is being
inlined. In a subsequent release, however, we are going to be
exposing accumulate_observations which may result in some
overhead, and will necessitate the change we're making here anyway
as part of the new API.