Remove AI calls to inform_shannon_pmi
At each time step of each initial condition of each time series,
inform_shannon_pmi was being called. Since inform_shannon_pmi is
not static it is not inlined, and because it does quite a lot of
error checking it is pretty costly. We've removed the call and
replaced it with a single call to log2 (and some memory accesses).