Remove argument and switch return order
The black_box_parts function took an argument for the length of
each time series. This isn't necessary as there are no history
or future lengths currently involved in that black boxing
process. We've remove that arguments. The `n` argument should
now be the total number of time steps in the time series, i.e.
the number of initial conditions times the number of time steps
per initial condition.

In addition, we switched the order of the returns. Before, the
first `nparts` elements were the base of each time series and the
remaining elements were the black boxed states. Now those are
reversed so that the last nparts elements are the bases.