Improve performance of inform_active_info for large k
The use of shannon_mi within inform_active_info makes the implementation
clear, but it sacrifices some performance as the history length grows.
This is a result of the fact that much of the state space is empty, so a
lot of effort is wasted computing logarithms that needn't be computed.